Rename build to build-eject, impl for build, shell commands
