+++ lib/esyBuildEjectCommand.js
@@ -2,8 +2,8 @@
  * @flow
  */
 
-import type {PackageDb} from '../lib/PackageDb';
-import type {MakeRawItem, MakeDefine, MakeRule} from '../lib/Makefile';
+import type {PackageDb} from './PackageDb';
+import type {MakeRawItem, MakeDefine, MakeRule} from './Makefile';
 
 const childProcess = require('child_process');
 const path = require('path');
@@ -12,9 +12,9 @@ const {
   traversePackageDb,
   collectTransitiveDependencies,
   collectDependencies,
-} = require('../lib/PackageDb');
-const PackageEnvironment = require('../lib/PackageEnvironment');
-const Makefile = require('../lib/Makefile');
+} = require('./PackageDb');
+const PackageEnvironment = require('./PackageEnvironment');
+const Makefile = require('./Makefile');
 
 const ESY_SANDBOX_REF = '$(ESY__SANDBOX)';
